Create 0_export_fids2mzML
This will export from the folder generated by Compass Explorer (starting from somewhen 2015). The name of the exported mzML file will be that of the analyte-ID. The analyte ID is taken from the info file hidden in the folder structure. An additional random number is created to account for possible duplicates...